Dedupe declaration of Wallet signing components
