+++ common/v2/features/InteractWithContracts/InteractWithContractsFlow.tsx
@@ -3,7 +3,7 @@ import { withRouter, RouteComponentProps } from 'react-router-dom';
 import styled from 'styled-components';
 
 import { translateRaw } from 'v2/translations';
-import { ExtendedContentPanel, Tabs } from 'v2/components';
+import { ExtendedContentPanel, Tabs, WALLET_STEPS } from 'v2/components';
 import { ROUTE_PATHS, DEFAULT_NETWORK } from 'v2/config';
 import { useStateReducer } from 'v2/utils';
 import { ITxReceipt, ISignedTx, Tab } from 'v2/types';
@@ -13,7 +13,6 @@ import { BREAK_POINTS } from 'v2/theme';
 import { interactWithContractsInitialState, InteractWithContractsFactory } from './stateFactory';
 import { Interact, InteractionReceipt } from './components';
 import { ABIItem, InteractWithContractState } from './types';
-import { WALLET_STEPS } from './helpers';
 import InteractionConfirm from './components/InteractionConfirm';
 
 const { SCREEN_XS } = BREAK_POINTS;
